+++ test/HttpHandlerTest.php
@@ -1,20 +1,20 @@
 <?php
 namespace DTS\eBaySDK\Test;
 
-use DTS\eBaySDK\Handler;
+use DTS\eBaySDK\HttpHandler;
 use GuzzleHttp\Client;
 use GuzzleHttp\Handler\MockHandler;
 use GuzzleHttp\Psr7;
 use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 
-class HandlerTest extends \PHPUnit_Framework_TestCase
+class HttpHandlerTest extends \PHPUnit_Framework_TestCase
 {
     public function testWorksWithSuccessfulRequest()
     {
         $mock = new MockHandler([new Response(200, [], 'OK')]);
         $client = new Client(['handler' => $mock]);
-        $httpHandler = new Handler($client);
+        $httpHandler = new HttpHandler($client);
 
         $request = new Request('POST', 'http://example.com', [], '');
         $response = $httpHandler($request)->wait()->getBody()->getContents();
@@ -25,7 +25,7 @@ class HandlerTest extends \PHPUnit_Framework_TestCase
     {
         $mock = new MockHandler([new \Exception('FAIL')]);
         $client = new Client(['handler' => $mock]);
-        $httpHandler = new Handler($client);
+        $httpHandler = new HttpHandler($client);
 
         $request = new Request('POST', 'http://example.com', [], '');
         try {