+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * The protected method `\DTS\eBaySDK\Services::BaseService::callOperation` has been removed. Code that called this method will now need to call `callOperationAsync(...)->wait()` instead.
 * The `handler` configuration option is now called [httpHandler](http://devbay.net/sdk/guides/guide/configuration.html#httpHandler).
+* The class `\DTS\eBaySDK\Handler` is now `\DTS\eBaySDK\HttpHandler`.
 
 ## 3.0.0 - 2016-04-11